Appease MSVC after the switch to more standards-compliant options.

+++ src/ui/Windows/ImgStatic.cpp
@@ -289,7 +289,7 @@ void CImgStatic::DrawItem(LPDRAWITEMSTRUCT lpDrawItemStruct)
 
     if (m_iZoomFactor == 10) {
       // Draw it
-      grp.DrawImage(&image, rc.left, rc.top, rc.right - rc.left, rc.bottom - rc.top);
+      grp.DrawImage(&image, static_cast<INT>(rc.left), static_cast<INT>(rc.top), static_cast<INT>(rc.right - rc.left), static_cast<INT>(rc.bottom - rc.top));
     } else {
       Gdiplus::Rect rcDest(rc.left, rc.top, rc.right - rc.left, rc.bottom - rc.top);
       int srcx, srcy, srcwidth, srcheight;
@@ -321,7 +321,7 @@ BOOL CImgStatic::OnEraseBkgnd(CDC *pDC)
     gdipColor.SetFromCOLORREF(GetSysColor(COLOR_3DFACE));
 
     Gdiplus::SolidBrush brush(gdipColor);
-    grp.FillRectangle(&brush, rc.left, rc.top, rc.right - rc.left, rc.bottom - rc.top);
+    grp.FillRectangle(&brush, static_cast<INT>(rc.left), static_cast<INT>(rc.top), static_cast<INT>(rc.right - rc.left), static_cast<INT>(rc.bottom - rc.top));
     grp.Flush();
 
     return TRUE;
@@ -357,7 +357,7 @@ void CImgStatic::ClearImage()
   gdipColor.SetFromCOLORREF(GetSysColor(COLOR_3DFACE));
 
   Gdiplus::SolidBrush brush(gdipColor);
-  grp.FillRectangle(&brush, rc.left, rc.top, rc.right - rc.left, rc.bottom - rc.top);
+  grp.FillRectangle(&brush, static_cast<INT>(rc.left), static_cast<INT>(rc.top), static_cast<INT>(rc.right - rc.left), static_cast<INT>(rc.bottom - rc.top));
   grp.Flush();
 }
